exceptionhandling changed - BadEnvironment exception introduced to tackle environment exceptions - exceptions isolated in the exceptions.py file - modifications made to the other file to import the exceptions
