+++ oandapy/__init__.py
@@ -1,2 +1,3 @@
-from .oandapy import API, OandaError
+from .oandapy import API
+from .exceptions import OandaError, BadEnvironment
 from .stream.stream import Streamer
