+++ oandapy/stream/stream.py
@@ -2,7 +2,7 @@ import warnings
 
 import json
 import requests
-from exceptions import BadEnvironment
+from ..exceptions import BadEnvironment
 from abc import ABCMeta, abstractmethod
 
 """ OANDA API wrapper for OANDA's REST API """
